renaming evaluation_strategies to strategies and improving README
